implementation css first work and images error solved

+++ src/main/java/com/example/semestrovka2/service/UploadFileService.java
@@ -1,34 +1,36 @@
 package com.example.semestrovka2.service;
 
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-
+import org.springframework.web.multipart.MultipartFile;
 @Service
 public class UploadFileService {
-    private String folder = "images//";
+    private final String folder = "src/main/resources/static/images/";
 
     public String saveImage(MultipartFile file) throws IOException {
         if (!file.isEmpty()) {
             byte[] bytes = file.getBytes();
             Path path = Paths.get(folder + file.getOriginalFilename());
+            // Ensure the directory exists
+            if (Files.notExists(path.getParent())) {
+                Files.createDirectories(path.getParent());
+            }
             Files.write(path, bytes);
             return file.getOriginalFilename();
-
         }
         return "default.jpg";
     }
-    public void deleteImage(String nombre){
-        String ruta="images//";
-        File file = new File(ruta+nombre);
-        file.delete();
 
+    public void deleteImage(String nombre) {
+        String ruta = "src/main/resources/static/images/";
+        File file = new File(ruta + nombre);
+        if (file.exists()) {
+            file.delete();
+        }
     }
 }
-
-
